usb: typec: anx7688: Properly handle system suspend
Shut down the periodic timer. Remove the work from the workqueue
and wait for it to finish. Re-check the status of the anx7688 after
resume and restart the timer.

I did some suspend/resume cycle testing while plugging and removing
the PD charger at random times, and the driver seems to do the right
thing, now.

Signed-off-by: Ondrej Jirman <megous@megous.com>